:boom: BREAKING CHANGE: Add fill method.
Renamed `nextFloat` to `nextFloat64`. See #1.

+++ src/nextFloat64.js
@@ -1,4 +1,4 @@
-export default function nextFloat(prng) {
+export default function nextFloat64(prng) {
 	const a = prng.next().value;
 	const b = prng.next().value;
 	return (a >>> 0) / 2 ** 32 + (b >>> 0) / 2 ** 64;
